Conflicts:
	frontend/views/site/contact.php

+++ simpeg/controllers/KenaikanGajiController.php
@@ -58,7 +58,7 @@ class KenaikanGajiController extends Controller {
      * If creation is successful, the browser will be redirected to the 'view' page.
      */
     public function actionGetPegawai() {
-        if (!empty($_POST['bulan']) and !empty($_POST['tahun'])) {
+        if (!empty($_POST['bulan']) and ! empty($_POST['tahun'])) {
             $bulan = substr("0" . $_POST['bulan'], -2, 2);
             $query = KenaikanGaji::model()->findAll(array('condition' => 'month(tmt_baru) = "' . $bulan . '" and year(tmt_baru) = "' . $_POST['tahun'] . '"'));
             echo $this->renderPartial('/kenaikanGaji/_tablePegawai', array('query' => $query, 'bulan' => $bulan, 'tahun' => $_POST['tahun']));
@@ -67,16 +67,15 @@ class KenaikanGajiController extends Controller {
 
     public function actionGetListPegawai() {
         if (!empty($_POST['unit_kerja'])) {
-              if (!empty($_POST['bulan']) and !empty($_POST['tahun'])) {
+            if (!empty($_POST['bulan']) and ! empty($_POST['tahun'])) {
                 $bulan = substr("0" . $_POST['bulan'], -2, 2);
                 logs($_POST);
 
-                $query = Pegawai::model()->with('Pangkat','JabatanStruktural')->findAll(array('condition' => 't.kedudukan_id = 1 and JabatanStruktural.unit_kerja_id = '.$_POST['unit_kerja'].' and month(t.tmt_cpns) <= "' . $bulan . '" ', 'order' => 'Pangkat.golongan_id ASC'));
+                $query = Pegawai::model()->with('Pangkat', 'JabatanStruktural')->findAll(array('condition' => 't.kedudukan_id = 1 and JabatanStruktural.unit_kerja_id = ' . $_POST['unit_kerja'] . ' and month(t.tmt_cpns) <= "' . $bulan . '" ', 'order' => 'Pangkat.golongan_id ASC'));
                 echo $this->renderPartial('/kenaikanGaji/_tableListPegawai', array('query' => $query, 'bulan' => $bulan, 'tahun' => $_POST['tahun']));
             }
-           
         } else {
-            if (!empty($_POST['bulan']) and !empty($_POST['tahun'])) {
+            if (!empty($_POST['bulan']) and ! empty($_POST['tahun'])) {
                 $bulan = substr("0" . $_POST['bulan'], -2, 2);
                 logs($_POST);
 
@@ -94,39 +93,41 @@ class KenaikanGajiController extends Controller {
 
         if (isset($_POST['proses'])) {
             for ($i = 0; $i < count($_POST['pegawai_id']); $i++) {
-                $bulan = substr("0" . $_POST['bulan'], -2, 2);
-                $valPegawai = Pegawai::model()->findByPk($_POST['pegawai_id'][$i]);
+                if (isset($_POST['dibayar'][$i])) {
+                    $bulan = substr("0" . $_POST['bulan'], -2, 2);
+                    $valPegawai = Pegawai::model()->findByPk($_POST['pegawai_id'][$i]);
 
-                $tglKenaikan = date("Y-m-d", strtotime($_POST['tmt_mulai'][$i]));
+                    $tglKenaikan = date("Y-m-d", strtotime($_POST['tmt_mulai'][$i]));
 
-                $delRiwayatgaji = RiwayatGaji::model()->deleteAll(array('condition' => 'pegawai_id=' . $valPegawai->id . ' and month(tmt_mulai)="' . $bulan . '" and year(tmt_mulai)="' . $_POST['tahun'] . '"'));
-                $riwayatGaji = new RiwayatGaji;
-                $riwayatGaji->nomor_register = date("ymisd");
-                $riwayatGaji->pegawai_id = $valPegawai->id;
-                $riwayatGaji->gaji = $_POST['gaji_baru'][$i];
-                $riwayatGaji->dasar_perubahan = "Kenaikan gaji berkala bulan " . $bulan . " tahun " . $_POST['tahun'];
-                $riwayatGaji->tmt_mulai = $_POST['tmt_mulai'][$i];
-                $riwayatGaji->save();
+                    $delRiwayatgaji = RiwayatGaji::model()->deleteAll(array('condition' => 'pegawai_id=' . $valPegawai->id . ' and month(tmt_mulai)="' . $bulan . '" and year(tmt_mulai)="' . $_POST['tahun'] . '"'));
+                    $riwayatGaji = new RiwayatGaji;
+                    $riwayatGaji->nomor_register = date("ymisd");
+                    $riwayatGaji->pegawai_id = $valPegawai->id;
+                    $riwayatGaji->gaji = $_POST['gaji_baru'][$i];
+                    $riwayatGaji->dasar_perubahan = "Kenaikan gaji berkala bulan " . $bulan . " tahun " . $_POST['tahun'];
+                    $riwayatGaji->tmt_mulai = $_POST['tmt_mulai'][$i];
+                    $riwayatGaji->save();
 
-                $delKgb = KenaikanGaji::model()->deleteAll(array('condition' => 'pegawai_id=' . $valPegawai->id . ' and month(tanggal)="' . $bulan . '" and year(tanggal)="' . $_POST['tahun'] . '"'));
-                $kgb = new KenaikanGaji;
-                $kgb->pegawai_id = $valPegawai->id;
-                $kgb->gaji_pokok_lama = $_POST['gaji_lama'][$i];
-                $kgb->gaji_pokok_baru = $_POST['gaji_baru'][$i];
-                $kgb->tmt_lama = $_POST['tmt_lama'][$i];
-                $kgb->tmt_baru = $_POST['tmt_mulai'][$i];
-                $kgb->created = date("Y-m-d h:i:s");
-                $kgb->nomor_register = '';
-                $kgb->sifat = '';
-                $kgb->perihal = '';
-                $kgb->pejabat = '';
-                $kgb->tanggal = $_POST['tmt_mulai'][$i];
-                $kgb->no_sk_akhiir = $_POST['no_sk_akhir'][$i];
-                $kgb->tanggal_sk_akhir = $_POST['tanggal_sk_akhir'][$i];
-                $kgb->save();
+                    $delKgb = KenaikanGaji::model()->deleteAll(array('condition' => 'pegawai_id=' . $valPegawai->id . ' and month(tanggal)="' . $bulan . '" and year(tanggal)="' . $_POST['tahun'] . '"'));
+                    $kgb = new KenaikanGaji;
+                    $kgb->pegawai_id = $valPegawai->id;
+                    $kgb->gaji_pokok_lama = $_POST['gaji_lama'][$i];
+                    $kgb->gaji_pokok_baru = $_POST['gaji_baru'][$i];
+                    $kgb->tmt_lama = $_POST['tmt_lama'][$i];
+                    $kgb->tmt_baru = $_POST['tmt_mulai'][$i];
+                    $kgb->created = date("Y-m-d h:i:s");
+                    $kgb->nomor_register = '';
+                    $kgb->sifat = '';
+                    $kgb->perihal = '';
+                    $kgb->pejabat = '';
+                    $kgb->tanggal = $_POST['tmt_mulai'][$i];
+                    $kgb->no_sk_akhiir = $_POST['no_sk_akhir'][$i];
+                    $kgb->tanggal_sk_akhir = $_POST['tanggal_sk_akhir'][$i];
+                    $kgb->save();
 
-                $valPegawai->riwayat_gaji_id = $riwayatGaji->id;
-                $valPegawai->save();
+                    $valPegawai->riwayat_gaji_id = $riwayatGaji->id;
+                    $valPegawai->save();
+                }
             }
         }
 
@@ -170,8 +171,7 @@ class KenaikanGajiController extends Controller {
             // if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
             if (!isset($_GET['ajax']))
                 $this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
-        }
-        else
+        } else
             throw new CHttpException(400, 'Invalid request. Please do not repeat this request again.');
     }
 
@@ -207,12 +207,12 @@ class KenaikanGajiController extends Controller {
             Yii::app()->end();
         }
     }
-    
-    public function actionGenerateExcel(){
+
+    public function actionGenerateExcel() {
         $tahun = $_GET['tahun'];
         $bulan = $_GET['bulan'];
         $unit_kerja = $_GET['unit_kerja'];
-        
+
         $criteria = new CDbCriteria;
         if (!empty($nama))
             $criteria->compare('nama', $nama, true);
@@ -234,7 +234,7 @@ class KenaikanGajiController extends Controller {
             $criteria->compare('jabatan_fu_id', $jabatan_fu_id);
         if (!empty($kode))
             $criteria->compare('kode', $kode);
-        
+
 //        $criteria->addCondition('kode IN (20,40)');
 
         $model = Honorer::model()->findAll($criteria);
@@ -244,7 +244,6 @@ class KenaikanGajiController extends Controller {
                     'model' => $model
                         ), true)
         );
-
     }
 
 }
